isYarnClusterMode extracts to LabelUtil
